+++ app/cli.py
@@ -4,13 +4,12 @@ app.cli
 
 Define app's commandline functions.
 """
+
 import click
 from flask import Flask
 from flask.cli import with_appcontext
 
-
-def command() -> None:
-    """Placeholder for future create funcs."""
+from .user import create_user_cli
 
 
 @click.group()
@@ -18,11 +17,11 @@ def create() -> None:
     """App init methods."""
 
 
-@create.command("command")
+@create.command("user")
 @with_appcontext
-def create_command() -> None:
-    """Placeholder for future create commands."""
-    command()
+def create_user() -> None:
+    """Create a new user with cli instead of forms."""
+    create_user_cli()
 
 
 def init_app(app: Flask) -> None: